Moved column name data from Instance to Hierarchy
